+++ packages/components/atoms/f-form-field/src/components/FormDropdown.vue
@@ -24,20 +24,20 @@
                 :value="option.value">{{ option.text }}</option>
         </select>
 
-        <caret-icon
+        <caret-down-filled-icon
             :class="$style['c-formDropdown-icon']"
             :data-test-id="testId.icon" />
     </div>
 </template>
 
 <script>
-import { CaretIcon } from '@justeat/f-vue-icons';
+import { CaretDownFilledIcon } from '@justeattakeaway/pie-icons-vue';
 
 export default {
     name: 'FormDropdown',
 
     components: {
-        CaretIcon
+        CaretDownFilledIcon
     },
 
     inheritAttrs: false,