+++ packages/components/atoms/f-form-field/CHANGELOG.md
@@ -3,6 +3,15 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+v4.12.0
+------------------------------
+*Jun 7, 2022*
+
+### Changed
+- Swapped CaretIcon with Fixed up CaretDownFilledIcon via @justeattakeaway/pie-icons-vue
+
+
 v4.11.0
 ------------------------------
 *May 30, 2022*